backend/app/core/security.py
from datetime import datetime, timedelta, timezone
import hashlib
import secrets

import jwt
from passlib.context import CryptContext

from app.core.config import settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def hash_password(password: str) -> str:
    return pwd_context.hash(password)


def verify_password(password: str, password_hash: str) -> bool:
    return pwd_context.verify(password, password_hash)


def create_access_token(user_id: str, email: str, role: str) -> str:
    exp = datetime.now(timezone.utc) + timedelta(minutes=settings.jwt_expire_minutes)
    payload = {"sub": user_id, "email": email, "role": role, "exp": exp}
    return jwt.encode(payload, settings.jwt_secret, algorithm="HS256")


def decode_token(token: str) -> dict:
    return jwt.decode(token, settings.jwt_secret, algorithms=["HS256"])


def random_token(length: int = 24) -> str:
    return secrets.token_urlsafe(length)[:length]


def stable_storage_key(file_id: str, filename: str) -> str:
    ext = filename.split(".")[-1] if "." in filename else "bin"
    digest = hashlib.sha256(file_id.encode()).hexdigest()[:16]
    return f"{file_id}_{digest}.{ext}"
