backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
minio==7.2.12
